Refactor user authentication and context management; update imports, implement user registration service, and enhance user model with proper export syntax.
