+++ client/src/App.jsx
@@ -1,8 +1,7 @@
 import { useState } from 'react'
 import LoginPage from './Pages/login'
 import './App.css'
-import { UserContextProvider } from 'context/UserContext'
-
+import { UserContextProvider} from '@/context/UserContext'
 function App() {
   return (
     <>
