+++ client/src/context/UserContext.jsx
@@ -1,4 +1,6 @@
-const {createContext, useContext, useState } = require("react");
+import { createContext, useState } from "react"
+
+
 
 const UserContext = createContext()
 const UserContextProvider = ({children})=>{
