+++ server/src/app.js
@@ -21,7 +21,7 @@ app.use((req,res,next)=>{
     next();
 })
 
-import {AuthRouter} from "./routes/UserAuth.routes"
+import AuthRouter  from "./routes/UserAuth.routes.js"
 
 app.use('api/v1/auth',AuthRouter)
 export {app}
